Refactor syscall random
