+++ ckb-debugger/src/arch_unix.rs
@@ -2,6 +2,10 @@ pub fn println(s: &str) {
     println!("{}", s);
 }
 
+pub fn random() -> u64 {
+    rand::random()
+}
+
 pub fn timestamp() -> u64 {
     std::time::SystemTime::now().duration_since(std::time::SystemTime::UNIX_EPOCH).unwrap().as_secs()
 }
